+++ ui/litellm-dashboard/src/app/(dashboard)/usage/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import UsagePageView from "@/components/Usage/components/UsagePageView";
+import UsagePageView from "@/components/UsagePage/components/UsagePageView";
 import useAuthorized from "@/app/(dashboard)/hooks/useAuthorized";
 import useTeams from "@/app/(dashboard)/hooks/useTeams";
