+++ ui/litellm-dashboard/src/components/activity_metrics.tsx
@@ -1,10 +1,10 @@
 import React from "react";
 import { Card, Grid, Text, Title } from "@tremor/react";
 import { AreaChart, BarChart } from "@tremor/react";
-import { DailyData, ModelActivityData, KeyMetricWithMetadata, TopApiKeyData } from "./Usage/types";
+import { DailyData, ModelActivityData, KeyMetricWithMetadata, TopApiKeyData } from "./UsagePage/types";
 import { Collapse } from "antd";
 import { formatNumberWithCommas } from "@/utils/dataUtils";
-import { valueFormatter } from "./Usage/utils/value_formatters";
+import { valueFormatter } from "./UsagePage/utils/value_formatters";
 import { CustomTooltip, CustomLegend } from "./common_components/chartUtils";
 
 interface ActivityMetricsProps {
