+++ ui/litellm-dashboard/src/components/common_components/chartUtils.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import type { CustomTooltipProps } from "@tremor/react";
-import { SpendMetrics } from "../Usage/types";
+import { SpendMetrics } from "../UsagePage/types";
 
 interface ChartDataPoint {
   date: string;
